move KEY_ENDUSER_PSEUDO_ID to a better constant file location
