+++ src/api-users/provider/constants/index.ts
@@ -1,0 +1,1 @@
+export const KEY_ENDUSER_PSEUDO_ID = 'enduser.pseudo.id';
