+++ src/api-users/provider/types.ts
@@ -1,4 +1,4 @@
-import { KEY_ENDUSER_PSEUDO_ID } from '../../constants/attributes.js';
+import { KEY_ENDUSER_PSEUDO_ID } from './constants/index.js';
 
 export interface User {
   [KEY_ENDUSER_PSEUDO_ID]: string;
