+++ src/constants/attributes.ts
@@ -2,7 +2,6 @@ export const KEY_EMB_TYPE = 'emb.type';
 export const KEY_EMB_STATE = 'emb.state';
 // TODO: update once we have the right type
 export const KEY_JS_EXCEPTION_STACKTRACE = 'emb.stacktrace.rn';
-export const KEY_ENDUSER_PSEUDO_ID = 'enduser.pseudo.id';
 
 export enum EMB_TYPES {
   Session = 'ux.session',
