+++ src/instrumentations/user/LocalStorageUserInstrumentation/types.ts
@@ -1,5 +1,5 @@
 import { User } from '../../../api-users/provider/types.js';
-import { KEY_ENDUSER_PSEUDO_ID } from '../../../constants/attributes.js';
+import { KEY_ENDUSER_PSEUDO_ID } from '../../../api-users/provider/constants/index.js';
 
 export const isUser = (user: unknown | User): user is User =>
   typeof (user as User)[KEY_ENDUSER_PSEUDO_ID] === 'string' &&
